Add a URL handler for raw commands.
Add a URL handler for raw commands that just forwards commands (with
no additional parameters) to a node. For example, using the URL
/nodes/node/1/command/raw/space.server.objects.count would execute the
command 'space.server.objects.count' on the server and report the raw
result back. This is proof of concept for interacting with
nodes. These URLs aren't linked anywhere; you have to enter commands
manually.