Add object sessions page with disconnect buttons.
Add object sessions page and link to it from the node info
page. Refactor the request made in the raw command page to be a
reusable method for issuing commands and add support for POST data as
well as the command name. Add a disconnect URL handler which just
makes the disconnection request and redirects to the sessions page.