Add commands page listing and linking all raw commands a node supports.
Add a new page, linked from node pages, that lists all available
commands and links them to the raw page for that command so we can
easily get at commands even if we don't have special support for
them. This also required adding '-' support to the URLs for raw
commands. Since this page requires reversing URLs, it'll only work if
the URL regex supports all commands listed by the node.